solve circular import issue by combining and renaming
