+++ tests/unit/test_web_app.py
@@ -17,11 +17,10 @@ import requests
 
 sys.path.append(".")
 import web_app.utils  # isort:skip  # noqa # pylint: disable=E0401,C0413
-from web_app.data_source import (  # isort:skip  # noqa # pylint: disable=E0401,C0413
+from web_app.networking import (  # isort:skip  # noqa # pylint: disable=E0401,C0413
     data_source as src,
     table_config as tc,
-    utils,
-    institution_info,
+    connections,
 )
 
 WBS = "mo"
@@ -31,9 +30,9 @@ WBS = "mo"
 def clear_all_cachetools_func_caches() -> Iterator[None]:
     """Clear all `cachetools.func` caches, everywhere."""
     yield
-    institution_info._cached_get_institutions_infos.cache_clear()  # type: ignore[attr-defined]
+    connections._cached_get_institutions_infos.cache_clear()  # type: ignore[attr-defined]
     tc.TableConfigParser._cached_get_configs.cache_clear()  # type: ignore[attr-defined]
-    web_app.utils.oidc_tools.CurrentUser._cached_get_info.cache_clear()  # type: ignore[attr-defined]
+    web_app.networking.connections.CurrentUser._cached_get_info.cache_clear()  # type: ignore[attr-defined]
 
 
 @pytest.fixture
@@ -55,7 +54,7 @@ def tconfig() -> Iterator[tc.TableConfigParser]:
         }
     }
     with patch(
-        "web_app.data_source.table_config.TableConfigParser._cached_get_configs"
+        "web_app.networking.table_config.TableConfigParser._cached_get_configs"
     ) as mock_cgc:
         mock_cgc.return_value = tconfig_cache
         yield tc.TableConfigParser(WBS)
@@ -219,7 +218,7 @@ class TestDataSource:
     @pytest.fixture
     def mock_rest(mocker: Any) -> Any:
         """Patch mock_rest."""
-        return mocker.patch("web_app.data_source.utils._rest_connection")
+        return mocker.patch("web_app.networking.connections._rest_connection")
 
     @staticmethod
     def test_pull_data_table(mock_rest: Any, tconfig: tc.TableConfigParser) -> None:
@@ -267,12 +266,12 @@ class TestDataSource:
             assert ret == response["table"]
 
     @staticmethod
-    @patch("web_app.utils.oidc_tools.CurrentUser._get_info")
+    @patch("web_app.networking.connections.CurrentUser._get_info")
     def test_push_record(
         current_user: Any, mock_rest: Any, tconfig: tc.TableConfigParser
     ) -> None:
         """Test push_record()."""
-        current_user.return_value = web_app.utils.oidc_tools.UserInfo(
+        current_user.return_value = web_app.networking.connections.UserInfo(
             "t.hanks", ["/institutions/IceCube/UW-Madison/_admin"]
         )
         unrealistic_hardcoded_resp = {
@@ -325,10 +324,10 @@ class TestDataSource:
             assert ret == unrealistic_hardcoded_resp["record"]
 
     @staticmethod
-    @patch("web_app.utils.oidc_tools.CurrentUser._get_info")
+    @patch("web_app.networking.connections.CurrentUser._get_info")
     def test_delete_record(current_user: Any, mock_rest: Any) -> None:
         """Test delete_record()."""
-        current_user.return_value = web_app.utils.oidc_tools.UserInfo(
+        current_user.return_value = web_app.networking.connections.UserInfo(
             "t.hanks", ["/institutions/IceCube/UW-Madison/_admin"], "foobarbaz"
         )
         record_id = "23"
@@ -354,13 +353,13 @@ class TestDataSource:
         )
 
     @staticmethod
-    @patch("web_app.utils.oidc_tools.CurrentUser.is_loggedin")
-    @patch("web_app.utils.oidc_tools.CurrentUser._get_info")
+    @patch("web_app.networking.connections.CurrentUser.is_loggedin")
+    @patch("web_app.networking.connections.CurrentUser._get_info")
     def test_list_snapshot_timestamps(
         current_user: Any, mock_ili: Any, mock_rest: Any
     ) -> None:
         """Test list_snapshot_timestamps()."""
-        current_user.return_value = web_app.utils.oidc_tools.UserInfo(
+        current_user.return_value = web_app.networking.connections.UserInfo(
             "t.hanks", ["/tokens/mou-dashboard-admin"], "foobarbaz"
         )
         mock_ili.return_value = True
@@ -384,10 +383,10 @@ class TestDataSource:
         assert sorted(ret, key=lambda k: k["timestamp"]) == response["snapshots"]
 
     @staticmethod
-    @patch("web_app.utils.oidc_tools.CurrentUser._get_info")
+    @patch("web_app.networking.connections.CurrentUser._get_info")
     def test_create_snapshot(current_user: Any, mock_rest: Any) -> None:
         """Test create_snapshot()."""
-        current_user.return_value = web_app.utils.oidc_tools.UserInfo(
+        current_user.return_value = web_app.networking.connections.UserInfo(
             "t.hanks", ["/institutions/IceCube/UW-Madison/_admin"], "foobarbaz"
         )
         response = {"timestamp": "a", "foo": "bar"}
@@ -412,7 +411,7 @@ class TestTableConfig:
     @pytest.fixture
     def mock_rest(mocker: Any) -> Any:
         """Patch mock_rest."""
-        return mocker.patch("web_app.data_source.utils._rest_connection")
+        return mocker.patch("web_app.networking.connections._rest_connection")
 
     @staticmethod
     def test_consts(tconfig: tc.TableConfigParser) -> None:
