+++ web_app/contents/wbs_generic_admin_callbacks.py
@@ -10,13 +10,12 @@ import dash_core_components as dcc  # type: ignore[import]
 from dash.dependencies import Input, Output, State  # type: ignore[import]
 
 from ..config import app
-from ..data_source import data_source as src
-from ..data_source import institution_info
-from ..data_source import table_config as tc
-from ..data_source.utils import DataSourceException
+from ..networking import connections
+from ..networking import data_source as src
+from ..networking import table_config as tc
+from ..networking.connections import CurrentUser, DataSourceException
 from ..utils import dash_utils as du
 from ..utils import types, utils
-from ..utils.oidc_tools import CurrentUser
 
 _CHANGES_COL: Final[str] = "Changes"
 
@@ -171,7 +170,7 @@ def summarize(
     except DataSourceException:
         return [], []
 
-    insts_infos = institution_info.get_institutions_infos()
+    insts_infos = connections.get_institutions_infos()
 
     column_names = ["Institution", "Institutional Lead"]
     if wbs_l1 == "mo":
