+++ web_app/networking/connections.py
@@ -1,15 +1,82 @@
-"""Handle user log-in and account info."""
+"""Utilities for MoU REST & OIDC interfaces."""
 
+
+import copy
 import logging
 import re
 from dataclasses import dataclass
-from typing import Any, Dict, List, Optional, cast
+from typing import Any, Dict, Final, List, Optional, cast
 
 import cachetools.func  # type: ignore[import]
 import flask  # type: ignore[import]
+import requests
 
-from ..config import MAX_CACHE_MINS, oidc
-from ..data_source import institution_info
+# local imports
+from rest_tools.client import RestClient  # type: ignore
+
+from ..config import MAX_CACHE_MINS, get_config_vars, oidc
+
+
+class DataSourceException(Exception):
+    """Exception class for bad data-source requests."""
+
+
+def _rest_connection() -> RestClient:
+    """Return REST Client connection object."""
+    config_vars = get_config_vars()
+    user_access_token = CurrentUser.get_access_token()
+
+    rc = RestClient(
+        config_vars["REST_SERVER_URL"], token=user_access_token, timeout=30, retries=0
+    )
+
+    return rc
+
+
+def _get_log_body(method: str, url: str, body: Any) -> str:
+    log_body = body
+
+    if method == "POST" and url.startswith("/table/data/"):
+        log_body = copy.deepcopy(body)
+        length: Final[int] = 10
+        b64 = log_body["base64_file"]
+        omitted = len(b64) - length
+        log_body["base64_file"] = f"{b64[:length]}... ({omitted} chars omitted)"
+
+    return str(log_body)
+
+
+def mou_request(method: str, url: str, body: Any = None) -> Dict[str, Any]:
+    """Make a request to the MoU REST server."""
+    log_body = _get_log_body(method, url, body)
+    logging.info(f"REQUEST :: {method} @ {url}, body: {log_body}")
+
+    try:
+        response: Dict[str, Any] = _rest_connection().request_seq(method, url, body)
+    except requests.exceptions.HTTPError as e:
+        logging.exception(f"EXCEPTED: {e}")
+        raise DataSourceException(str(e))
+
+    def log_it(key: str, val: Any) -> Any:
+        if key == "table":
+            return f"{len(val)} records"
+        if isinstance(val, dict):
+            return val.keys()
+        return val
+
+    logging.info(f"RESPONSE ({method} @ {url}, body: {log_body}) ::")
+    for key, val in response.items():
+        logging.debug(f"> {key}")
+        logging.debug(f"-> {str(type(val).__name__)}")
+        logging.debug(f"-> {log_it(key, val)}")
+
+    return response
+
+
+#
+# OIDC Tools
+# Handle user log-in and account info.
+#
 
 
 @dataclass(frozen=True)
@@ -108,7 +175,7 @@ class CurrentUser:
                 group_insts.add(m.groupdict()["inst"])
 
         # now, check if each of the institutions has an mou
-        infos = institution_info.get_institutions_infos()
+        infos = get_institutions_infos()
         editable_insts = []
         for short_name in group_insts:
             if not infos[short_name].has_mou:
@@ -124,3 +191,31 @@ class CurrentUser:
     def get_access_token() -> str:
         """Retrieve the logged-in user's access token."""
         return CurrentUser._get_info().access_token
+
+
+#
+# Institution Info
+#
+
+
+@dataclass(frozen=True)
+class Institution:
+    """Hold minimal institution data."""
+
+    short_name: str
+    long_name: str
+    is_us: bool
+    has_mou: bool
+    institution_lead_uid: str
+
+
+@cachetools.func.ttl_cache(ttl=MAX_CACHE_MINS * 60)  # type: ignore[misc]
+def _cached_get_institutions_infos() -> Dict[str, Institution]:
+    logging.warning("Cache Miss: _cached_get_institutions_infos()")
+    resp = cast(Dict[str, Dict[str, Any]], mou_request("GET", "/institution/today"))
+    return {k: Institution(**v) for k, v in resp.items()}
+
+
+def get_institutions_infos() -> Dict[str, Institution]:
+    """Get a dict of all institutions with their info."""
+    return cast(Dict[str, Institution], _cached_get_institutions_infos())
