+++ web_app/networking/data_source.py
@@ -3,10 +3,9 @@
 
 from typing import Any, Dict, Final, List, Optional, Tuple, TypedDict, Union, cast
 
+from ..networking.connections import CurrentUser, mou_request
 from ..utils import types, utils
-from ..utils.oidc_tools import CurrentUser
 from . import table_config as tc
-from .utils import mou_request
 
 # constants
 _OC_SUFFIX: Final[str] = "_original"
